feat: making authorized requests

+++ src/components/Workout.tsx
@@ -3,6 +3,7 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 type Props = {
     workout: IWorkout;
@@ -10,9 +11,16 @@ type Props = {
 
 export default function Workout({ workout }: Props) {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
+
     const handleClick = async () => {
+        if (!user) return;
+
         const res = await fetch(`${import.meta.env.VITE_API_URL}/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user?.token}`
+            }
         });
         const json = await res.json();
